projects/algo_hack/backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
sse-starlette>=1.8.2
google-generativeai>=0.5.0
httpx>=0.27.0
python-dotenv>=1.0.0
algosdk>=2.6.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
